make add to cart button funcitonal on see details
